fix: stripe integration landing page
