+++ pages/account/onboarding/index.js
@@ -1,5 +1,7 @@
 import { authOptions } from "../../api/auth/[...nextauth]";
 import { getServerSession } from "next-auth/next";
+import { useRouter } from "next/router";
+import { useSession } from "next-auth/react";
 import {
   FaGithub,
   FaLink,
@@ -19,6 +21,7 @@ import Card from "@components/Card";
 import Button from "@components/Button";
 import Navigation from "@components/account/manage/Navigation";
 import ProgressBar from "@components/statistics/ProgressBar";
+import Alert from "@components/Alert";
 
 export async function getServerSideProps(context) {
   const session = await getServerSession(context.req, context.res, authOptions);
@@ -73,6 +76,22 @@ export async function getServerSideProps(context) {
 }
 
 export default function Onboarding({ profile, progress }) {
+  const router = useRouter();
+  const { data: session } = useSession();
+  if (typeof window !== "undefined" && window.localStorage) {
+    if (router.query.alert) {
+      localStorage.removeItem("premium-intent");
+    }
+    if (
+      session &&
+      session.accountType !== "premium" &&
+      localStorage.getItem("premium-intent")
+    ) {
+      localStorage.removeItem("premium-intent");
+      router.push("/api/stripe");
+    }
+  }
+
   const cards = [
     {
       icon: FaPersonBurst,
@@ -135,6 +154,11 @@ export default function Onboarding({ profile, progress }) {
       isEdit: profile.milestones && profile.milestones.length > 0,
     },
   ];
+
+  const alerts = {
+    premium: "You are now a premium user!",
+    cancel: "You cancelled your subscription.",
+  };
   return (
     <>
       <PageHead
@@ -144,6 +168,11 @@ export default function Onboarding({ profile, progress }) {
 
       <Page>
         <Navigation />
+
+        {router.query.alert && (
+          <Alert type="info" message={alerts[router.query.alert]} />
+        )}
+
         <div className="flex flex-col mb-8 md:flex-row">
           <h1 className="mb-4 text-4xl font-bold grow">
             Create &amp; Manage Your Profile
