+++ pages/api/stripe.js
@@ -45,8 +45,8 @@ export default async function handler(req, res) {
     ],
     mode: "subscription",
     customer: session.stripeCustomerId,
-    success_url: `${clientEnv.NEXT_PUBLIC_BASE_URL}/account/statistics?alert=premium`,
-    cancel_url: `${clientEnv.NEXT_PUBLIC_BASE_URL}/account/statistics?alert=cancel`,
+    success_url: `${clientEnv.NEXT_PUBLIC_BASE_URL}/account/onboarding?alert=premium`,
+    cancel_url: `${clientEnv.NEXT_PUBLIC_BASE_URL}/account/onboarding?alert=cancel`,
     payment_method_collection: "if_required",
     allow_promotion_codes: true,
     ...subscription,
